pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weld-traceability-agent"
version = "0.1.0"
description = "Telegram-first agent layer for weld traceability workflows."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "openai>=2.11.0",
  "pydantic>=2.9",
  "PyYAML>=6.0",
  "python-telegram-bot>=21.7",
  "weld-traceability-assistant @ git+https://github.com/wycffff/Weld_map_understanding_and_weld_traceability_assistant.git@fc2549696ba7fe3bfa858b93520ef954b2127b28",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
]

[project.scripts]
weld-agent = "weld_traceability_agent.cli:main"
weld-agent-telegram = "weld_traceability_agent.telegram.bot:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
